add json-c support for rhel9 build rule.

+++ work/build.sh
@@ -101,7 +101,7 @@ case "${SYSTEM_ID}" in
     dnf group install -y "Development Tools"
     dnf install -y kernel-devel kernel-rpm-macros kernel-abi-stablelists
     dnf install -y which kmod
-    dnf install -y krb5-devel keyutils-libs-devel openssl-devel libyaml-devel libmount-devel libnl3-devel
+    dnf install -y krb5-devel keyutils-libs-devel openssl-devel libyaml-devel libmount-devel libnl3-devel json-c-devel
 
     KERNEL_VERSION="$(rpm -qa kernel-devel | sed -e 's+kernel-devel-++')"
     KERNEL_ARCH="$(echo ${KERNEL_VERSION} | sed -e 's+.*\.++g')"
@@ -109,14 +109,14 @@ case "${SYSTEM_ID}" in
     [ ! -d ${DEST_DIR} ] && mkdir -p ${DEST_DIR}
     cd ${DEST_DIR}
 
-    #if [ -f "${LUSTRE_SOURCE}" ]; then
-    #  tar zxvf "${LUSTRE_SOURCE}"
-    #  pushd $(basename "${LUSTRE_SOURCE}" .tar.gz)
-    #else
+    if [ -f "${LUSTRE_SOURCE}" ]; then
+      tar zxvf "${LUSTRE_SOURCE}"
+      pushd $(basename "${LUSTRE_SOURCE}" .tar.gz)
+    else
       dnf install -y git
       git clone --branch b2_15 --depth 1 git://git.whamcloud.com/fs/lustre-release.git
       pushd lustre-release
-    #fi
+    fi
     sh ./autogen.sh
     ./configure --disable-server --disable-lru-resize --with-linux=/usr/src/kernels/${KERNEL_VERSION}
     make srpm
@@ -128,7 +128,7 @@ case "${SYSTEM_ID}" in
     MOCK_CONFIG="$(echo $SYSTEM_ID | sed -e 's+centos+centos-stream+g' -e 's+9\..*$+9+g')-$KERNEL_ARCH"
     mock -r ${MOCK_CONFIG} init
     mock -r ${MOCK_CONFIG} install kernel-devel-${KERNEL_VERSION} kernel-headers-${KERNEL_VERSION} kernel-abi-stablelists
-    mock -r ${MOCK_CONFIG} install krb5-devel keyutils-libs-devel openssl-devel libyaml-devel libmount-devel libnl3-devel
+    mock -r ${MOCK_CONFIG} install krb5-devel keyutils-libs-devel openssl-devel libyaml-devel libmount-devel libnl3-devel json-c-devel
     mock -r ${MOCK_CONFIG} --no-clean --define "configure_args '--disable-lru-resize'" --define "kdir /usr/src/kernels/${KERNEL_VERSION}" --define "kobjdir /usr/src/kernels/${KERNEL_VERSION}" --without servers --without ldiskfs --with gss_keyring --with gss --with gss --without snmp ${SOURCE_RPM}
     cp /var/lib/mock/${MOCK_CONFIG}/result/*.rpm ${DEST_DIR}/
     cp /var/lib/mock/${MOCK_CONFIG}/result/build.log ${DEST_DIR}/
